coord: colocate session vars and var-related types
The old organization didn't make a lot of sense. Colocate the actual
session variables with the types involved in declaration session
variables.
